Add new TAC instructions and improve bytecode translation
This commit adds the following three-address code instructions to
the compiler backend along with bytecode translation support:

- AddressOf
- Load
- Store
- LessThan

This commit also implements translation support for the Call
instruction, adds typed variables to indicate whether a variable is
a parameter, local, or global, and cleans up the translation code
in general.

The AutoGen helper class is also updated to support object
construction via a user-supplied function.