Implement the main menu
This commit also changes the menu sequence to account for the main
menu. The welcome screen is no longer accessible after the main menu
is entered and the character select menu now returns to the previous
screen when closed. The PlainMenu class has also been modified to
draw a larger menu background.