Improve error handling for expected lvalues
This commit adds code to raise an exception when a non-lvalue is
used in an expression that requires an lvalue. A PlusExpr class for
unary plus expressions is also added to prevent unary plus results
from being used as lvalues.

This commit also adds a temporary check for valid function calls.