Added the win condition
The user wins the level if, after the code finishes running, the
Robot is not moving, is on the Door tile, and has the same color
as the Door. If the user fails, the level is reset, but the user's
code is left untouched in the buffer.

Other changes needed to implement this:
  * Merge the DoorR and DoorG into the Door class
  * Give the Door class a color property that changes the tile
  * Change the level object dictionary to create a Door with a
    given color if DoorR or DoorG is specified in the level *.csv
    file.