Cleanup/reorganization attempt
LevelScreen is now LevelSplashScreen

LevelScreen now contains the text editor, the level, and the control
logic for playing the game.

The new Screen class should be the base class for all screens (menus,
game screens, ...). The time argument is removed from draw() and
state updates are now handled by an update(time) function.