Add unreachable code removal to the C code generator
This commit adds unreachable code removal to the C code generator to
reduce the size of the resulting bytecode. The code for unreachable
statements is still generated to allow for error checking in this
stage, but it is not added to the final output. If error checking is
moved to an earlier stage then this temporary code generation will no
longer be necessary.