Add a debug window to the level screen
This commit adds a debug window to the level screen that can be
accessed by pressing F8. The debug window displays the processor's
memory and registers.

This commit also changes the way Robot handles processor lifetime.
The processor is no longer destroyed when an error occurs. A running
flag is used instead to determine whether the step function should
run the processor.

The size of the processor's memory is also changed from 0x1000 (4K)
to 0x400 (1K) words to make it easier to view all of the memory.