Add a setup script for executable generation
The setup script uses py2exe to create a Windows executable.

Asset paths are changed to paths below the working directory to
account for the directory structure when running the generated
executable. A check is also added to main.py to update the working
directory as necessary to account for this.

include.h is moved into assets so that it remains accessible after
the executable is generated.

The pickle icon is used as the icon for the Windows executable.