Can complete level 101
The robot can now move while it is halted.

LevelScreens use the Robot's _pos variable instead of x,y when
displaying the variables to display status while the robot is
halted but still moving.

The level success condition allows the robot to be SUCCESS_EPSILON
(currently 0.5) away from the door when it stops.